Refactor WordBlur to another component
